Bugfix: Set colorscale for curvature plots comparable through plots
The colorscale (colorrange) for the curvature plots was recalculated for
every contour segment. This made it impossible to compare the curvature
values between different plots. Now the curvature is calculated for
every frame first to find the max and min curvature values. Then these
values are used to to set the colorbar.